==> notebooks/3-results.ipynb <==
cd ..

import pandas as pd
import numpy as np 
from sacrebleu import sentence_bleu

## Result

### Pretrained

test_scores_en_da_sv_pretrained = pd.read_csv('data/kaggle/test_scores_en_da_sv_pretrained.csv')
test_scores_en_da_sv_pretrained['bleu'] = test_scores_en_da_sv_pretrained.apply(lambda x: sentence_bleu(x['translation'], [x['tgt_text']]).score, axis=1)
test_scores_en_da_sv_pretrained['bleu_lowercase'] = test_scores_en_da_sv_pretrained.apply(lambda x: sentence_bleu(x['translation'].lower(), 
                                                                                                                  [x['tgt_text'].lower()]).score,
                                                                                          axis=1)

test_scores_en_da_sv_pretrained.groupby(['src_lang', 'tgt_lang'])[['bleu', 'bleu_lowercase']].mean()

### Finetuned

en_sv_finetuned_results = pd.read_csv('data/kaggle/en_sv_finetuned_results.csv')

en_sv_finetuned_results['bleu'] = en_sv_finetuned_results.apply(lambda x: sentence_bleu(x['prediction'], [x['target_text']]).score, axis=1)
en_sv_finetuned_results['pretrained_bleu'] = en_sv_finetuned_results.apply(lambda x: sentence_bleu(x['pretrained_predictions'], [x['target_text']]).score, axis=1)
en_sv_finetuned_results['bleu_lowercase'] =  en_sv_finetuned_results.apply(lambda x: sentence_bleu(x['prediction'].lower(), [x['target_text'].lower()]).score, axis=1)
en_sv_finetuned_results['pretrained_bleu_lowercase'] = en_sv_finetuned_results.apply(lambda x: sentence_bleu(x['pretrained_predictions'].lower(), 
                                                                                                             [x['target_text'].lower()]).score, axis=1)


en_sv_finetuned_results.groupby(['src_lang', 'tgt_lang'])[['bleu', 'bleu_lowercase', 'pretrained_bleu', 'pretrained_bleu_lowercase']].mean()
